FIX: In FirstCycleIntensityPFLaneTest, the value in the message is an integer.

+++ src/main/java/fr/ens/transcriptome/aozan/tests/FirstCycleIntensityPFLaneTest.java
@@ -42,7 +42,7 @@ public class FirstCycleIntensityPFLaneTest extends AbstractSimpleLaneTest {
   @Override
   protected Class<?> getValueType() {
 
-    return Double.class;
+    return Integer.class;
   }
 
   @Override
